pyproject.toml
[project]
name = "system-monitor-agent"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.116.1",
    "google-adk>=1.9.0",
    "google-generativeai>=0.8.5",
    "python-dotenv>=1.1.1",
    "uvicorn>=0.35.0",
]
